Update hash algos based on Mozilla Root Store Policy
Mozilla CA policy:
https://www.mozilla.org/en-US/about/governance/policies/security-group/certs/policy/

When a root or intermediate certificate's ECDSA key is used to produce a signature, only the following algorithms may be used, and with the following encoding requirements:

If the signing key is P-256, the signature MUST use ECDSA with SHA-256. The encoded AlgorithmIdentifier MUST match the following hex-encoded bytes: 300a06082a8648ce3d040302.

If the signing key is P-384, the signature MUST use ECDSA with SHA-384. The encoded AlgorithmIdentifier MUST match the following hex-encoded bytes: 300a06082a8648ce3d040303.

Signed-off-by: Vesa Jääskeläinen <dachaac@gmail.com>